Fixed Search not working for Anonymous Users
With the new dj name search feature, non-authenticated users were unable to search bceause of RLS policies on the Profiles tables. As opposed to allowing all users access this data, i created a postgres  View with information that can be publically accessed. I lastly modified the code to query from that view instead of profiles.